UI: Change file structure to route groups
