+++ monkey/monkey_island/cc/next_ui/src/app/(protected)/home/page.tsx
@@ -1,0 +1,3 @@
+export default function HomePage() {
+    return <div>home page</div>;
+}
