+++ monkey/monkey_island/cc/next_ui/src/app/(protected)/layout.tsx
@@ -1,0 +1,6 @@
+import '../../styles/globals.scss';
+import MainLayout from '@/layouts/main-layout/MainLayout';
+
+export default function AppLayout({ children }: { children: React.ReactNode }) {
+    return <MainLayout>{children}</MainLayout>;
+}
