+++ monkey/monkey_island/cc/next_ui/src/app/(protected)/page.module.scss
@@ -1,0 +1,6 @@
+.test {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  min-height: 100vh;
+}
