+++ monkey/monkey_island/cc/next_ui/src/app/(protected)/page.tsx
@@ -1,0 +1,13 @@
+import styles from './page.module.scss';
+import Test from '@/_components/Test';
+
+export default function Home() {
+    return (
+        // <Test />
+        // <main className={styles.main}>
+        <div className={styles.test}>
+            <Test />
+        </div>
+        // </main>
+    );
+}
