+++ monkey/monkey_island/cc/next_ui/src/styles/globals.scss
@@ -1,0 +1,14 @@
+:root {
+
+}
+
+html,
+body {
+  max-width: 100vw;
+  overflow-x: hidden;
+}
+
+body {
+
+}
+
