+++ monkey/monkey_island/cc/next_ui/src/styles/variables.scss
@@ -1,1 +1,2 @@
 $border-radius-25: 25px;
+$app-bar-height: 50px;
